Fixed an event resetting issue in ODE.
When several discrete events occur during the same ODE integration step,
they are handled chronologically or reverse chronologically depending on
the integration direction. If one of the event truncates the step (for
example because its eventOccurred method returns RESET or
RESET_DERIVATIVES for example), the stepAccepted method of the pending
events later in the step were not called. This implied that in the next
step, these events were still referring to data from previous step, they
had lost synchronization with the integrator.

JIRA: MATH-695

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1189086 13f79535-47bb-0310-9956-ffa450edef68