Improved Dormand-Prince 8(5,3) step interpolator accuracy at step start.
The previous step is preserved and if interpolation time is in the first
half ot the step, then interpolation is based on step start, otherwise
is it based on step end. Previously, interpolation was always performed
with respect to step end.

If this trick proves useful, it will be extended to other Runge-Kutta
type step interpolators.

JIRA: MATH-705

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1204270 13f79535-47bb-0310-9956-ffa450edef68