Clarified origin of the class.
The original comments seemed to show the class was created from
Mixtools, which is GPL code. As per discussion on the JIRA issue
MATH-817 (see https://issues.apache.org/jira/browse/MATH-817), it
appears it is original work and NOT a derived work from Mixtools.
In fact, Mixtools was used mainly to verify the results and generate
some reference data (see the Junit test cases). Neither the code nor the
API from Mixtools were used.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1459876 13f79535-47bb-0310-9956-ffa450edef68