Improved reporting section in pom.xml.
The Apache RAT plugin was configured in the build section instead of the
reporting section, so it did not automatically used our local
configuration.

The checkstyle anf findbugs plugins were updated to recent versions,
with some adaptation to prevent double entries.

The CPD plugin has been temporarily disabled because in 3.X it detects
huge duplications, which is true but misleading, as the duplications are
a consequence of our upward compatibility policy.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1462745 13f79535-47bb-0310-9956-ffa450edef68