Don't advertise an assumption we may remove soon.
We will probably extend the circle class to also represent small circles
on the 2-sphere, not only great circles. This would introduce an
aperture angle that could be different from \pi/2. In this case, the
part of a circle inside another circle may exist or not (small circles
can be disjoint), and may have a length either shorter or longer than
\pi.

This change is not done yet, but could happen any time soon, even before
3.3 is out.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1557981 13f79535-47bb-0310-9956-ffa450edef68