Deprecated AVLTree and OrderedTuple.
These two classes were dirty hacks used only for the extraction of
boundary from 2D Euclidean BSP trees. They were not really efficient and
produces some problems as topology was not preserved.

A completely new implementation of boundary extraction has been set up
as of 3.4, that preserves more topological information and therefore
should be more robust. This new implementation does not use these
classes anymore.

As AVLTree is not really a math concept, this class is not considered to
belong to [math] scope. OrderedTuple is more a grey zone, but it is
really something inefficient and no need for it is foreseen.