Removed LAPACK license and notice.
Strangely enough, we do not include any code from LAPACK anymore since
several releases! The last reference was for eigen-decomposition and it
was changed long ago to another implementation.
The only material loosely related to LAPACK are some reference values in
tests which were computed using LAPACK, but these are output of LAPACK
routines, not code from LAPACK routines.