Attempt to circumvent some errors which seem to be platform-dependent.
The Jenkins build often fails on code that seems to be perfectly
correct. Failures also do no always happen so they may depend on
platform. There were similar problems a few months ago that were
probably related to JIT bugs.

This fix simply tries to do the same thing as before, but with an
earlier detection of NaN in one case, and by comparing directly the bits
representation in another case, to avoid wrong optimizations.