MATH-1419: Invalid usage of exception in PolynomialSplineFunction
Fix order of parameter for NumberIsTooSmallException when length of
knots is less than 2 in the constructor of PolynomialSplineFunction to
be fit with the definition of the exception. Moreover, boundedIsAllowed
value is changed to true since 2 is also allowed value.

Change the error message format for this exception appropriately in
English and French.

Since this does not affect any error during using the library, but to
make the code more concrete, following definition strictly is important
things to implement functionalities.