MATH-1371: Elkan's enhancement to "K-Means" algorithm.
Original implementation is a old PR provided by Artem Barger.
Updated and changed to pass the code style checks.

I did not review how this algorithm departs from the other
implementations available in the same package, and whether
some other design could minimize duplicate codes (but it
would have been a pity to drop a potential performance
improvement).
A thorough examination should be considered, when tackling
the refactoring of the "clustering" package.

Closes #35.