Use a LinkedHashMap to respect the order of the input data
Note: The int/real enumerated distributions do not specify that the
input must be sorted. This change supports the legacy behaviour enforced
by the unit test which is assuming input values are processed in
encounter order (which happens to be sorted).

Given int and double have a natural ordering it may be better to update
the distribution to sort input values. However this would contradict the
enumerated distribution which respects the input order of the list used
in the constructor.