Set to default strategies if null is given
In the current implementation, it is possible to set the `nanStrategy` and
`tiesStrategy` of `NaturalRanking` to `null`, even though they will
inevitably throw NPEs upon any calls to `rank` or `resolveTie`.
This commit modifies the constructor such that if `null` is given to either
of the strategies, then they are set to the default strategies.