MATH-1671: Update stat.descriptive package to use Commons Statistics
Refactor the SummaryStatistics default implementations to use
DoubleStatistics.

Removes method:

getPopulationVariance
getSecondMoment

The population variance relies the second moment implementation which
computes a statistic related to the central second moment. It is not a
standard statistic and is not supported in Commons Statistics.

Updates the min/max implementations to use Math.min/max. Previous
behaviour ignored NaN values. The change now matches with JDK stream
behaviour.